requirements.txt
m3u8
requests
pycryptodome
beautifulsoup4